Split out hotkey and utilities
